// src/generated/prisma/enums.ts
export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  RECRUITER: 'RECRUITER',
  CLIENT: 'CLIENT',
  CANDIDATE: 'CANDIDATE',
  EMPLOYEE: 'EMPLOYEE'
} as const

export type Role = (typeof Role)[keyof typeof Role]